Refactor;  - Separating the admin controller functionalities out from admin routes;
